backend/requirements.txt
Flask
flask-cors
pymongo
joblib
scikit-learn
xgboost
python-whois
dnspython
pyjwt
bcrypt
gunicorn
